+++ packages/class-core/src/index.ts
@@ -22,6 +22,7 @@ export * from "./deserializer.js";
 export { deepDuplicateWith } from "./duplicator.js";
 export * from "./id-mapping.js";
 export { combinedFactoryFor } from "./factory.js";
+export { Forest, ObservableForest } from "./forest.js";
 // skip linking.js: see comment there
 export * from "./LionCore_builtins.g.js";
 export type { IdOrUnresolved } from "./references.js";