+++ packages/delta-protocol-test/src/tests/scenarios-with-only-client.tests.ts
@@ -99,13 +99,13 @@ describe(`scenarios where TypeScript client is the only system under test (using
 
         expect(lionWebClient.participationId!).to.equal("participation-a")
 
-        const newPartition = lionWebClient.createNode(testLanguageBase.LinkTestConcept, "partition-A")
+        const newPartition = lionWebClient.forest.createNode(testLanguageBase.LinkTestConcept, "partition-A")
         lionWebClient.addPartition(newPartition)
-        expect(lionWebClient.model).to.deep.equal([newPartition])
+        expect(lionWebClient.forest.partitions).to.deep.equal([newPartition])
 
         // assert idempotency of adding a new partition:
         lionWebClient.addPartition(newPartition)
-        expect(lionWebClient.model).to.deep.equal([newPartition])
+        expect(lionWebClient.forest.partitions).to.deep.equal([newPartition])
 
         await lionWebClient.disconnect()
     })