+++ packages/delta-protocol-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.8.0 — not yet released
+
+* Rework the `LionWebClient` class to use the new `Forest` class from the `class-core` package.
+
+
 ## 0.7.1
 
 (The 0.7.0 release was deprecated because its `validation` package was faulty.)